graphql/requirements.txt
aniso8601==9.0.1
anyio==3.7.0
astunparse==1.6.3
backports.cached-property==1.0.2
click==8.1.3
exceptiongroup==1.1.1
fastapi==0.96.0
graphene==3.2.2
graphql-core==3.2.3
graphql-relay==3.2.0
grpcio==1.54.2
grpcio-tools==1.54.2
h11==0.14.0
idna==3.4
importlib-metadata==6.6.0
pika==1.3.2
Pillow==9.5.0
protobuf==4.23.2
psycopg2==2.9.6
pydantic==1.10.8
python-dateutil==2.8.2
python-gql==0.2.4
python-multipart==0.0.6
reportlab==4.0.4
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
starlette-graphene3==0.6.0
starlette-graphql==0.2.1
strawberry-graphql==0.183.6
typing-extensions==4.6.3
uvicorn==0.22.0
zipp==3.15.0
